Switch to Dagster for orchestration
