+++ pyproject.toml
@@ -8,10 +8,9 @@ package-mode = false
 [tool.poetry.dependencies]
 python = "^3.11"
 duckdb = "*"
-schedule = "*"
+dagster = "*"
 dbt-core = "*"
 dbt-duckdb = "*"
-prefect = "*"
 pandas = "*"
 yfinance = "*"
